separationg plots.h into different classes .h and .cpp
