+++ main/mainwindow.cc
@@ -11,7 +11,9 @@
 #include "../lib/data/error_model.h"
 #include "../lib/data/measurement_model.h"
 #include "../lib/data/titles_model.h"
-#include "../lib/plots/plots.h"
+#include "../lib/plots/heatmap2d.h"
+#include "../lib/plots/sorted_line_plot.h"
+#include "../lib/plots/unsorted_line_plot.h"
 #include "./ui_mainwindow.h"
 #include "dialogwindow.h"
 #include "qcustomplot.h"
